Massive restructuring (ID system, input events, tiles, direction).
After looking at some past projects, I realised I was doing it wrong (no wonder it felt so weird and complicated). Ditched the old direction system, created Entities and some basic Tiles, and started with an ID system that allows Entities to be identified by a specific Integer, that can be shared by the server to the client (including on the same machine).