Started adding Weapons and Rubble.
Rebuilt Tiles into an undirected-graph structure, where the removal of all the Tiles in a Tile's supportedBy set results in the destruction of that Tile, and the removal of the Tile from all of the Tiles it supports (hence the process goes on and on, enabling entire buildings to collapse).
Also started adding weapons - need to fill out the Weapon.singleBullet() method and the AssaultRifle class (needs constants!).