Ugh, what a mess. This is NOT a finished commit by any standard (simply backing up my work).
Attempted to figure out the structure of how the Round enactments will be done. It's not trivial! I think I will use sequences of events, as that is the most general and flexible solution. Still have lots to figure out about how I should implement subsequences and so on - or even how the movement should work (should it be instantaneous? I can't believe I didn't decide this earlier; that was a major oversight).
Also rearranged and further developed weapons, now approaching completion for the base types (will have an AssaultRifle, SubMachineGun, SniperRifle & Shotgun - not counting the RPG - may also have a couple of pistols and a knife for secondary weapons).
Also (again) made large progress on implementing Rubble - at least backend-wise. Just waiting on the damage for crushed Entities (i.e., the events system).